Changed sound levels into float

+++ displayTemperature/models.py
@@ -7,14 +7,14 @@ class Device(models.Model):
     latitude = models.FloatField(blank=True, null=True)
     longitude = models.FloatField(blank=True, null=True)
     address = models.CharField(max_length=100, blank=True, null=True)
-    lowest = models.IntegerField(default=0)
-    highest = models.IntegerField(default=0)
+    lowest = models.FloatField(default=0)
+    highest = models.FloatField(default=0)
     def __str__(self):
         return self.dev_eui
 
 class Message(models.Model):
     device = models.ForeignKey(Device, on_delete=models.CASCADE, default = "0000000000000000")
-    average = models.IntegerField(default=0)
+    average = models.FloatField(default=0)
     rcv_date = models.DateTimeField()
     def __int__(self):
         return self.average
